+++ src/cli/commands/anon-col/helpers/get-processor.ts
@@ -1,5 +1,5 @@
 import {EngineType} from '../../../../config/types';
-import {Processor} from '../../../../processors/base-processor/base-processor';
+import {Processor} from '../../../../processors/base-processor/processor';
 import {MariaDbProcessor} from '../../../../processors/mariadb/mariadb-processor';
 import {MongoProcessor} from '../../../../processors/mongo/mongo-processor';
 import {PostgresProcessor} from '../../../../processors/postgres/postgres-processor';
