+++ src/processors/base-processor/base-processor.ts
@@ -1,17 +1,3 @@
-import {Anonymizer} from '../../anonymizers/types';
-import {Config} from '../../config/types';
-
-export interface Processor {
-	processDb?(config: Config, dbName?: string): void;
-	processTable?(config: Config, tableName: string, dbName?: string): void;
-	processColumn(
-		tableName: string,
-		columnName: string,
-		anonymizer: Anonymizer,
-		dbName?: string,
-	): Promise<void>;
-}
-
 export class BaseProcessor {
 	constructor(protected readonly uri: string) {}
 }
