+++ src/processors/mongo/mongo-processor.ts
@@ -1,7 +1,8 @@
 import {Db, MongoClient} from 'mongodb';
 import {Anonymizer} from '../../anonymizers/types';
 import {debugLogger} from '../../services/loggers/debug-logger';
-import {BaseProcessor, Processor} from '../base-processor/base-processor';
+import {BaseProcessor} from '../base-processor/base-processor';
+import {Processor} from '../base-processor/processor';
 
 const logger = debugLogger.extend('mongo-processor');
 export class MongoProcessor extends BaseProcessor implements Processor {
