+++ src/processors/postgres/postgres-processor.ts
@@ -1,7 +1,8 @@
 import knex, {Knex} from 'knex';
 import {Anonymizer} from '../../anonymizers/types';
 import {debugLogger} from '../../services/loggers/debug-logger';
-import {BaseProcessor, Processor} from '../base-processor/base-processor';
+import {BaseProcessor} from '../base-processor/base-processor';
+import {Processor} from '../base-processor/processor';
 
 const logger = debugLogger.extend('postgress-processor');
